requirements.txt
Flask==3.1.0
slack-sdk==3.33.4
Werkzeug==3.1.3
gunicorn==23.0.0
glueops-helpers @ https://github.com/GlueOps/python-glueops-helpers-library/archive/refs/tags/v0.4.1.zip
